+++ src/effects/mod.rs
@@ -5,6 +5,7 @@ mod greyscale;
 pub use blur::Blur;
 pub use dither::{
     Color, Colour, DiffusionAlgorithm, DiffusionKernel, DiffusionScan, DiffusionTap,
-    ErrorDiffusion, OrderedDither, Palette, Threshold, ThresholdMap, ThresholdRotation,
+    ErrorDiffusion, NoiseAlgorithm, NoiseDither, OrderedDither, Palette, Threshold, ThresholdMap,
+    ThresholdRotation,
 };
 pub use greyscale::Greyscale;