+++ src/lib.rs
@@ -11,7 +11,8 @@ mod selection;
 
 pub use effects::{
     Blur, Color, Colour, DiffusionAlgorithm, DiffusionKernel, DiffusionScan, DiffusionTap,
-    ErrorDiffusion, Greyscale, OrderedDither, Palette, Threshold, ThresholdMap, ThresholdRotation,
+    ErrorDiffusion, Greyscale, NoiseAlgorithm, NoiseDither, OrderedDither, Palette, Threshold,
+    ThresholdMap, ThresholdRotation,
 };
 pub use io::{read, write};
 pub use pipeline::{Pipeline, PipelineStep};